requirements.txt
mysql-connector-python
wonderwords
Faker
numpy
kafka-python
pyspark
requests
pandas
seaborn